Module Lacaml

Binding to the BLAS and LAPACK libraries. You can make use of this library by referring to the corresponding module you need for your precision and number type:

open Lacaml.S
open Lacaml.D
open Lacaml.C
open Lacaml.Z

To use this library, you should be familiar with BLAS and LAPACK. The following quick reference guide for the BLAS and LAPACK quick reference may be useful to you. For the precise description of the functions, consult the man pages online or, if you installed them on your machine (if you use Linux, they should be in the packages of your distribution), read them with Emacs: M-x man (under Unix) or M-x woman (all systems).

Precision-dependent Modules

Real-valued Modules

module S : sig ... end

Single-precision real BLAS and LAPACK functions.

module D : sig ... end

Double-precision real BLAS and LAPACK functions.

Complex-valued Modules

module C : sig ... end

Single-precision complex BLAS and LAPACK functions.

module Z : sig ... end

Double-precision complex BLAS and LAPACK functions.

Precision-independent Module

module Common : sig ... end

Types and functions common to all precision dependent sub-modules.

Pretty-printing Module

module Io : sig ... end

Pretty-printing of vectors and matrices.

Utility Module

module Utils : sig ... end

Utility functions