Module Lacaml__.Impl2_D

BLAS-1 interface

val dot : ?⁠n:int -> ?⁠ofsx:int -> ?⁠incx:int -> Lacaml__.Float64.vec -> ?⁠ofsy:int -> ?⁠incy:int -> Lacaml__.Float64.vec -> float

dot ?n ?ofsx ?incx x ?ofsy ?incy y see BLAS documentation!

parameter n

default = greater n s.t. ofsx+(n-1)(abs incx) <= dim x

parameter ofsx

default = 1

parameter incx

default = 1

parameter ofsy

default = 1

parameter incy

default = 1

val asum : ?⁠n:int -> ?⁠ofsx:int -> ?⁠incx:int -> Lacaml__.Float64.vec -> float

asum ?n ?ofsx ?incx x see BLAS documentation!

parameter n

default = greater n s.t. ofsx+(n-1)(abs incx) <= dim x

parameter ofsx

default = 1

parameter incx

default = 1

BLAS-2 interface

val sbmv : ?⁠n:int -> ?⁠k:int -> ?⁠ofsy:int -> ?⁠incy:int -> ?⁠y:Lacaml__.Float64.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> ?⁠up:bool -> ?⁠alpha:float -> ?⁠beta:float -> ?⁠ofsx:int -> ?⁠incx:int -> Lacaml__.Float64.vec -> Lacaml__.Float64.vec

sbmv ?n ?k ?ofsy ?incy ?y ?ar ?ac a ?up ?alpha ?beta ?ofsx ?incx x see BLAS documentation!

returns

vector y, which is overwritten.

parameter n

default = number of available columns to the right of ac.

parameter k

default = number of available rows in matrix a - 1

parameter ofsy

default = 1

parameter incy

default = 1

parameter ar

default = 1

parameter ac

default = 1

parameter y

default = uninitialized vector of minimal length (see BLAS)

parameter up

default = true i.e., upper band of a is supplied

parameter alpha

default = 1.0

parameter beta

default = 0.0

parameter ofsx

default = 1

parameter incx

default = 1

val ger : ?⁠m:int -> ?⁠n:int -> ?⁠alpha:float -> ?⁠ofsx:int -> ?⁠incx:int -> Lacaml__.Float64.vec -> ?⁠ofsy:int -> ?⁠incy:int -> Lacaml__.Float64.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> unit

ger ?m ?n ?alpha ?ofsx ?incx x ?ofsy ?incy y n ?ar ?ac a see BLAS documentation!

parameter m

default = number of rows of a

parameter n

default = number of columns of a

parameter alpha

default = 1.0

parameter ofsx

default = 1

parameter incx

default = 1

parameter ofsy

default = 1

parameter incy

default = 1

parameter ar

default = 1

parameter ac

default = 1

val syr : ?⁠n:int -> ?⁠alpha:float -> ?⁠up:bool -> ?⁠ofsx:int -> ?⁠incx:int -> Lacaml__.Float64.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> unit

syr ?n ?alpha ?up ?ofsx ?incx x ?ar ?ac a see BLAS documentation!

parameter n

default = number of rows of a

parameter alpha

default = 1.0

parameter up

default = true i.e., upper triangle of a is supplied

parameter ofsx

default = 1

parameter incx

default = 1

parameter ar

default = 1

parameter ac

default = 1

LAPACK interface

Auxiliary routines

val lansy_min_lwork : int -> Lacaml.Common.norm4 -> int

lansy_min_lwork m norm

returns

the minimum length of the work array used by the lansy-function.

parameter norm

type of norm that will be computed by lansy

parameter n

the number of columns (and rows) in the matrix

val lansy : ?⁠n:int -> ?⁠up:bool -> ?⁠norm:Lacaml.Common.norm4 -> ?⁠work:Lacaml__.Float64.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> float

lansy ?norm ?up ?n ?ar ?ac ?work a see LAPACK documentation!

parameter norm

default = `O

parameter up

default = true (reference upper triangular part of a)

parameter n

default = number of columns of matrix a

parameter work

default = allocated work space for norm `I

val lamch : [ `E | `S | `B | `P | `N | `R | `M | `U | `L | `O ] -> float

lamch cmach see LAPACK documentation!

Linear equations (computational routines)

val orgqr_min_lwork : n:int -> int

orgqr_min_lwork ~n

returns

the minimum length of the work-array used by the orgqr-function if the matrix has n columns.

val orgqr_opt_lwork : ?⁠m:int -> ?⁠n:int -> ?⁠k:int -> tau:Lacaml__.Float64.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> int

orgqr_opt_lwork ?m ?n ?k ~tau ?ar ?ac a

returns

the optimum length of the work-array used by the orgqr-function given matrix a, optionally its logical dimensions m and n, and the number of reflectors k.

parameter m

default = available number of rows in matrix a

parameter n

default = available number of columns in matrix a

parameter k

default = available number of elements in vector tau

val orgqr : ?⁠m:int -> ?⁠n:int -> ?⁠k:int -> ?⁠work:Lacaml__.Float64.vec -> tau:Lacaml__.Float64.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> unit

orgqr ?m ?n ?k ?work ~tau ?ar ?ac a see LAPACK documentation!

parameter m

default = available number of rows in matrix a

parameter n

default = available number of columns in matrix a

parameter k

default = available number of elements in vector tau

val ormqr_opt_lwork : ?⁠side:Lacaml.Common.side -> ?⁠trans:Lacaml.Common.trans2 -> ?⁠m:int -> ?⁠n:int -> ?⁠k:int -> tau:Lacaml__.Float64.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> ?⁠cr:int -> ?⁠cc:int -> Lacaml__.Float64.mat -> int

ormqr_opt_lwork ?side ?trans ?m ?n ?k ~tau ?ar ?ac a ?cr ?cc c

returns

the optimum length of the work-array used by the ormqr-function given matrix a and b, optionally its logical dimensions m and n, and the number of reflectors k.

parameter m

default = available number of rows in matrix a

parameter n

default = available number of columns in matrix a

parameter k

default = available number of elements in vector tau

val ormqr : ?⁠side:Lacaml.Common.side -> ?⁠trans:Lacaml.Common.trans2 -> ?⁠m:int -> ?⁠n:int -> ?⁠k:int -> ?⁠work:Lacaml__.Float64.vec -> tau:Lacaml__.Float64.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> ?⁠cr:int -> ?⁠cc:int -> Lacaml__.Float64.mat -> unit

ormqr ?side ?trans ?m ?n ?k ?work ~tau ?ar ?ac a ?cr ?cc c see LAPACK documentation!

parameter side

default = `L

parameter trans

default = `N

parameter m

default = available number of rows in matrix a

parameter n

default = available number of columns in matrix a

parameter k

default = available number of elements in vector tau

val gecon_min_lwork : int -> int

gecon_min_lwork n

returns

the minimum length of the work array used by the gecon-function.

parameter n

the logical dimensions of the matrix given to the gecon-function

val gecon_min_liwork : int -> int

gecon_min_liwork n

returns

the minimum length of the iwork array used by the gecon-function.

parameter n

the logical dimensions of the matrix given to gecon-function

val gecon : ?⁠n:int -> ?⁠norm:Lacaml.Common.norm2 -> ?⁠anorm:float -> ?⁠work:Lacaml__.Float64.vec -> ?⁠iwork:Lacaml.Common.int32_vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> float

gecon ?n ?norm ?anorm ?work ?rwork ?ar ?ac a

returns

estimate of the reciprocal of the condition number of matrix a

parameter n

default = available number of columns of matrix a

parameter norm

default = 1-norm

parameter anorm

default = norm of the matrix a as returned by lange

parameter work

default = automatically allocated workspace

parameter iwork

default = automatically allocated workspace

parameter ar

default = 1

parameter ac

default = 1

val sycon_min_lwork : int -> int

sycon_min_lwork n

returns

the minimum length of the work array used by the sycon-function.

parameter n

the logical dimensions of the matrix given to the sycon-function

val sycon_min_liwork : int -> int

sycon_min_liwork n

returns

the minimum length of the iwork array used by the sycon-function.

parameter n

the logical dimensions of the matrix given to sycon-function

val sycon : ?⁠n:int -> ?⁠up:bool -> ?⁠ipiv:Lacaml.Common.int32_vec -> ?⁠anorm:float -> ?⁠work:Lacaml__.Float64.vec -> ?⁠iwork:Lacaml.Common.int32_vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> float

sycon ?n ?up ?ipiv ?anorm ?work ?iwork ?ar ?ac a

returns

estimate of the reciprocal of the condition number of symmetric matrix a

parameter n

default = available number of columns of matrix a

parameter up

default = upper triangle of the factorization of a is stored

parameter ipiv

default = vec of length n

parameter anorm

default = 1-norm of the matrix a as returned by lange

parameter work

default = automatically allocated workspace

parameter iwork

default = automatically allocated workspace

val pocon_min_lwork : int -> int

pocon_min_lwork n

returns

the minimum length of the work array used by the pocon-function.

parameter n

the logical dimensions of the matrix given to the pocon-function

val pocon_min_liwork : int -> int

pocon_min_liwork n

returns

the minimum length of the iwork array used by the pocon-function.

parameter n

the logical dimensions of the matrix given to pocon-function

val pocon : ?⁠n:int -> ?⁠up:bool -> ?⁠anorm:float -> ?⁠work:Lacaml__.Float64.vec -> ?⁠iwork:Lacaml.Common.int32_vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> float

pocon ?n ?up ?anorm ?work ?iwork ?ar ?ac a

returns

estimate of the reciprocal of the condition number of symmetric positive definite matrix a

parameter n

default = available number of columns of matrix a

parameter up

default = upper triangle of Cholesky factorization of a is stored

parameter work

default = automatically allocated workspace

parameter iwork

default = automatically allocated workspace

parameter anorm

default = 1-norm of the matrix a as returned by lange

Least squares (expert drivers)

val gelsy_min_lwork : m:int -> n:int -> nrhs:int -> int

gelsy_min_lwork ~m ~n ~nrhs

returns

the minimum length of the work-array used by the gelsy-function if the logical dimensions of the matrix are m rows and n columns and if there are nrhs right hand side vectors.

val gelsy_opt_lwork : ?⁠m:int -> ?⁠n:int -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> ?⁠nrhs:int -> ?⁠br:int -> ?⁠bc:int -> Lacaml__.Float64.mat -> int

gelsy_opt_lwork ?m ?n ?ar ?ac a ?nrhs ?br ?bc b

returns

the optimum length of the work-array used by the gelsy-function given matrix a, optionally its logical dimensions m and n and given right hand side matrix b with an optional number nrhs of vectors.

parameter m

default = available number of rows in matrix a

parameter n

default = available number of columns in matrix a

parameter nrhs

default = available number of columns in matrix b

val gelsy : ?⁠m:int -> ?⁠n:int -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> ?⁠rcond:float -> ?⁠jpvt:Lacaml.Common.int32_vec -> ?⁠work:Lacaml__.Float64.vec -> ?⁠nrhs:int -> ?⁠br:int -> ?⁠bc:int -> Lacaml__.Float64.mat -> int

gelsy ?m ?n ?ar ?ac a ?rcond ?jpvt ?ofswork ?work ?nrhs b see LAPACK documentation!

returns

the effective rank of a.

parameter m

default = available number of rows in matrix a

parameter n

default = available number of columns of matrix a

parameter rcond

default = (-1) => machine precision

parameter jpvt

default = vec of length n

parameter work

default = vec of optimum length (-> gelsy_opt_lwork)

parameter nrhs

default = available number of columns in matrix b

val gelsd_min_lwork : m:int -> n:int -> nrhs:int -> int

gelsd_min_lwork ~m ~n ~nrhs

returns

the minimum length of the work-array used by the gelsd-function if the logical dimensions of the matrix are m and n and if there are nrhs right hand side vectors.

val gelsd_opt_lwork : ?⁠m:int -> ?⁠n:int -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> ?⁠nrhs:int -> ?⁠br:int -> ?⁠bc:int -> Lacaml__.Float64.mat -> int

gelsd_opt_lwork ?m ?n ?ar ?ac a ?nrhs b

returns

the optimum length of the work-array used by the gelsd-function given matrix a, optionally its logical dimensions m and n and given right hand side matrix b with an optional number nrhs of vectors.

parameter m

default = available number of rows in matrix a

parameter n

default = available number of columns in matrix a

parameter nrhs

default = available number of columns in matrix b

val gelsd_min_iwork : int -> int -> int

gelsd_min_iwork m n

returns

the minimum (= optimum) length of the iwork-array used by the gelsd-function if the logical dimensions of the matrix are m and n.

val gelsd : ?⁠m:int -> ?⁠n:int -> ?⁠rcond:float -> ?⁠ofss:int -> ?⁠s:Lacaml__.Float64.vec -> ?⁠work:Lacaml__.Float64.vec -> ?⁠iwork:Lacaml__.Float64.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> ?⁠nrhs:int -> ?⁠br:int -> ?⁠bc:int -> Lacaml__.Float64.mat -> int

gelsd ?m ?n ?rcond ?ofss ?s ?ofswork ?work ?ar ?ac a ?nrhs b see LAPACK documentation!

returns

the effective rank of a.

raises Failure

if the function fails to converge.

parameter m

default = available number of rows in matrix a

parameter n

default = available number of columns of matrix a

parameter rcond

default = (-1) => machine precision

parameter ofss

default = 1 or ignored if s is not given

parameter s

default = vec of length min rows cols

parameter work

default = vec of optimum length (-> gelsd_opt_lwork)

parameter iwork

default = vec of optimum (= minimum) length

parameter nrhs

default = available number of columns in matrix b

val gelss_min_lwork : m:int -> n:int -> nrhs:int -> int

gelss_min_lwork ~m ~n ~nrhs

returns

the minimum length of the work-array used by the gelss-function if the logical dimensions of the matrix are m rows and n columns and if there are nrhs right hand side vectors.

val gelss_opt_lwork : ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> ?⁠m:int -> ?⁠n:int -> ?⁠nrhs:int -> ?⁠br:int -> ?⁠bc:int -> Lacaml__.Float64.mat -> int

gelss_opt_lwork ?ar ?ac a ?m ?n ?nrhs ?br ?bc b

returns

the optimum length of the work-array used by the gelss-function given matrix a, optionally its logical dimensions m and n and given right hand side matrix b with an optional number nrhs of vectors.

parameter m

default = available number of rows in matrix a

parameter n

default = available number of columns in matrix a

parameter nrhs

default = available number of columns in matrix b

val gelss : ?⁠m:int -> ?⁠n:int -> ?⁠rcond:float -> ?⁠ofss:int -> ?⁠s:Lacaml__.Float64.vec -> ?⁠work:Lacaml__.Float64.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> ?⁠nrhs:int -> ?⁠br:int -> ?⁠bc:int -> Lacaml__.Float64.mat -> int

gelss ?m ?n ?rcond ?ofss ?s ?ofswork ?work ?ar ?ac a ?nrhs ?br ?bc b see LAPACK documentation!

returns

the effective rank of a.

raises Failure

if the function fails to converge.

parameter m

default = available number of rows in matrix a

parameter n

default = available number of columns of matrix a

parameter rcond

default = (-1) => machine precision

parameter ofss

default = 1 or ignored if s is not given

parameter s

default = vec of length min m n

parameter work

default = vec of optimum length (-> gelss_opt_lwork)

parameter nrhs

default = available number of columns in matrix b

General Schur factorization

val gees : ?⁠n:int -> ?⁠jobvs:Lacaml.Common.schur_vectors -> ?⁠sort:Lacaml.Common.eigen_value_sort -> ?⁠wr:Lacaml__.Float64.vec -> ?⁠wi:Lacaml__.Float64.vec -> ?⁠vsr:int -> ?⁠vsc:int -> ?⁠vs:Lacaml__.Float64.mat -> ?⁠work:Lacaml__.Float64.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> int * Lacaml__.Float64.vec * Lacaml__.Float64.vec * Lacaml__.Float64.mat

gees ?n ?jobvs ?sort ?w ?vsr ?vsc ?vs ?work ?ar ?ac a See gees-function for details about arguments.

returns

(sdim, wr, wi, vs)

General SVD routines

val gesvd_min_lwork : m:int -> n:int -> int

gesvd_min_lwork ~m ~n

returns

the minimum length of the work array used by the gesvd-function for matrices with m rows and n columns.

val gesvd_opt_lwork : ?⁠m:int -> ?⁠n:int -> ?⁠jobu:Lacaml.Common.svd_job -> ?⁠jobvt:Lacaml.Common.svd_job -> ?⁠s:Lacaml__.Float64.vec -> ?⁠ur:int -> ?⁠uc:int -> ?⁠u:Lacaml__.Float64.mat -> ?⁠vtr:int -> ?⁠vtc:int -> ?⁠vt:Lacaml__.Float64.mat -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> int

val gesvd : ?⁠m:int -> ?⁠n:int -> ?⁠jobu:Lacaml.Common.svd_job -> ?⁠jobvt:Lacaml.Common.svd_job -> ?⁠s:Lacaml__.Float64.vec -> ?⁠ur:int -> ?⁠uc:int -> ?⁠u:Lacaml__.Float64.mat -> ?⁠vtr:int -> ?⁠vtc:int -> ?⁠vt:Lacaml__.Float64.mat -> ?⁠work:Lacaml__.Float64.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> Lacaml__.Float64.vec * Lacaml__.Float64.mat * Lacaml__.Float64.mat

val gesdd_liwork : m:int -> n:int -> int

val gesdd_min_lwork : ?⁠jobz:Lacaml.Common.svd_job -> m:int -> n:int -> unit -> int

gesdd_min_lwork ?jobz ~m ~n

returns

the minimum length of the work array used by the gesdd-function for matrices with m rows and n columns for SVD-job jobz.

val gesdd_opt_lwork : ?⁠m:int -> ?⁠n:int -> ?⁠jobz:Lacaml.Common.svd_job -> ?⁠s:Lacaml__.Float64.vec -> ?⁠ur:int -> ?⁠uc:int -> ?⁠u:Lacaml__.Float64.mat -> ?⁠vtr:int -> ?⁠vtc:int -> ?⁠vt:Lacaml__.Float64.mat -> ?⁠iwork:Lacaml.Common.int32_vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> int

val gesdd : ?⁠m:int -> ?⁠n:int -> ?⁠jobz:Lacaml.Common.svd_job -> ?⁠s:Lacaml__.Float64.vec -> ?⁠ur:int -> ?⁠uc:int -> ?⁠u:Lacaml__.Float64.mat -> ?⁠vtr:int -> ?⁠vtc:int -> ?⁠vt:Lacaml__.Float64.mat -> ?⁠work:Lacaml__.Float64.vec -> ?⁠iwork:Lacaml.Common.int32_vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> Lacaml__.Float64.vec * Lacaml__.Float64.mat * Lacaml__.Float64.mat

General eigenvalue problem (simple drivers)

val geev_min_lwork : ?⁠vectors:bool -> int -> int

geev_min_lwork vectors n

returns

the minimum length of the work array used by the geev-function. vectors indicates whether eigenvectors are supposed to be computed.

parameter n

the logical dimensions of the matrix given to geev-function

parameter vectors

default = true

val geev_opt_lwork : ?⁠n:int -> ?⁠vlr:int -> ?⁠vlc:int -> ?⁠vl:Lacaml__.Float64.mat option -> ?⁠vrr:int -> ?⁠vrc:int -> ?⁠vr:Lacaml__.Float64.mat option -> ?⁠ofswr:int -> ?⁠wr:Lacaml__.Float64.vec -> ?⁠ofswi:int -> ?⁠wi:Lacaml__.Float64.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> int

geev_opt_lwork ?n ?vlr ?vlc ?vl ?vrr ?vrc ?vr ?ofswr wr ?ofswi wi ?ar ?ac a See geev-function for details about arguments.

returns

"optimal" size of work array.

val geev : ?⁠n:int -> ?⁠work:Lacaml__.Float64.vec -> ?⁠vlr:int -> ?⁠vlc:int -> ?⁠vl:Lacaml__.Float64.mat option -> ?⁠vrr:int -> ?⁠vrc:int -> ?⁠vr:Lacaml__.Float64.mat option -> ?⁠ofswr:int -> ?⁠wr:Lacaml__.Float64.vec -> ?⁠ofswi:int -> ?⁠wi:Lacaml__.Float64.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> Lacaml__.Float64.mat * Lacaml__.Float64.vec * Lacaml__.Float64.vec * Lacaml__.Float64.mat

geev ?work ?n ?vlr ?vlc ?vl ?vrr ?vrc ?vr ?ofswr ?wr ?ofswi ?wi ?ar ?ac a

returns

(lv, wr, wi, rv), where wr and wv are the real and imaginary components of the eigenvalues, and lv and rv are the left and right eigenvectors. lv (rv) is the empty matrix if vl (vr) is set to None.

raises Failure

if the function fails to converge

parameter n

default = available number of columns of matrix a

parameter work

default = automatically allocated workspace

parameter vl

default = Automatically allocated left eigenvectors. Pass None if you do not want to compute them, Some lv if you want to provide the storage. You can set vlr, vlc in the last case. (See LAPACK GEEV docs for details about storage of complex eigenvectors)

parameter vr

default = Automatically allocated right eigenvectors. Pass None if you do not want to compute them, Some rv if you want to provide the storage. You can set vrr, vrc in the last case.

parameter wr

default = vector of size n; real components of the eigenvalues

parameter wi

default = vector of size n; imaginary components of the eigenvalues

parameter a

the matrix whose eigensystem is computed

Symmetric-matrix eigenvalue and singular value problems (simple drivers)

val syev_min_lwork : int -> int

syev_min_lwork n

returns

the minimum length of the work-array used by the syev-function if the logical dimensions of the matrix are n.

val syev_opt_lwork : ?⁠n:int -> ?⁠vectors:bool -> ?⁠up:bool -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> int

syev_opt_lwork ?n ?vectors ?up ?ar ?ac a

returns

the optimum length of the work-array used by the syev-function given matrix a, optionally its logical dimension n and whether the eigenvectors must be computed (vectors).

parameter n

default = available number of columns of matrix a

parameter vectors

default = false, i.e. eigenvectors are not computed

parameter up

default = true, i.e. upper triangle of a is stored

val syev : ?⁠n:int -> ?⁠vectors:bool -> ?⁠up:bool -> ?⁠work:Lacaml__.Float64.vec -> ?⁠ofsw:int -> ?⁠w:Lacaml__.Float64.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> Lacaml__.Float64.vec

syev ?n ?vectors ?up ?ofswork ?work ?ofsw ?w ?ar ?ac a computes all eigenvalues and, optionally, eigenvectors of the real symmetric matrix a.

returns

the vector w of eigenvalues in ascending order.

raises Failure

if the function fails to converge.

parameter n

default = available number of columns of matrix a

parameter vectors

default = false i.e, eigenvectors are not computed

parameter up

default = true i.e., upper triangle of a is stored

parameter work

default = vec of optimum length (-> syev_opt_lwork)

parameter ofsw

default = 1 or ignored if w is not given

parameter w

default = vec of length n

val syevd_min_lwork : vectors:bool -> int -> int

syevd_min_lwork vectors n

returns

the minimum length of the work-array used by the syevd-function if the logical dimensions of the matrix are n and given whether eigenvectors should be computed (vectors).

val syevd_min_liwork : vectors:bool -> int -> int

syevd_min_liwork vectors n

returns

the minimum length of the iwork-array used by the syevd-function if the logical dimensions of the matrix are n and given whether eigenvectors should be computed (vectors).

val syevd_opt_lwork : ?⁠n:int -> ?⁠vectors:bool -> ?⁠up:bool -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> int

syevd_opt_lwork ?n ?vectors ?up ?ar ?ac a

returns

the optimum length of the work-array used by the syevd-function given matrix a, optionally its logical dimension n and whether the eigenvectors must be computed (vectors).

parameter n

default = available number of columns of matrix a

parameter vectors

default = false, i.e. eigenvectors are not computed

parameter up

default = true, i.e. upper triangle of a is stored

val syevd_opt_liwork : ?⁠n:int -> ?⁠vectors:bool -> ?⁠up:bool -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> int

syevd_opt_liwork ?n ?vectors ?up ?ar ?ac a

returns

the optimum length of the iwork-array used by the syevd-function given matrix a, optionally its logical dimension n and whether the eigenvectors must be computed (vectors).

parameter n

default = available number of columns of matrix a

parameter vectors

default = false, i.e. eigenvectors are not computed

parameter up

default = true, i.e. upper triangle of a is stored

val syevd_opt_l_li_work : ?⁠n:int -> ?⁠vectors:bool -> ?⁠up:bool -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> int * int

syevd_opt_l_li_iwork ?n ?vectors ?up ?ar ?ac a

returns

the tuple of optimum lengths of the work- and iwork-arrays respectively, used by the syevd-function given matrix a, optionally its logical dimension n and whether the eigenvectors must be computed (vectors).

parameter n

default = available number of columns of matrix a

parameter vectors

default = false, i.e. eigenvectors are not computed

parameter up

default = true, i.e. upper triangle of a is stored

val syevd : ?⁠n:int -> ?⁠vectors:bool -> ?⁠up:bool -> ?⁠work:Lacaml__.Float64.vec -> ?⁠iwork:Lacaml.Common.int32_vec -> ?⁠ofsw:int -> ?⁠w:Lacaml__.Float64.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> Lacaml__.Float64.vec

syevd ?n ?vectors ?up ?ofswork ?work ?iwork ?ofsw ?w ?ar ?ac a computes all eigenvalues and, optionally, eigenvectors of the real symmetric matrix a. If eigenvectors are desired, it uses a divide and conquer algorithm.

returns

the vector w of eigenvalues in ascending order.

raises Failure

if the function fails to converge.

parameter n

default = available number of columns of matrix a

parameter vectors

default = false i.e, eigenvectors are not computed

parameter up

default = true i.e., upper triangle of a is stored

parameter work

default = vec of optimum length (-> syev_opt_lwork)

parameter iwork

default = int32_vec of optimum length (-> syevd_opt_liwork)

parameter ofsw

default = 1 or ignored if w is not given

parameter w

default = vec of length n

val sbev_min_lwork : int -> int

sbev_min_lwork n

returns

the minimum length of the work-array used by the sbev-function if the logical dimensions of the matrix are n.

val sbev : ?⁠n:int -> ?⁠kd:int -> ?⁠zr:int -> ?⁠zc:int -> ?⁠z:Lacaml__.Float64.mat -> ?⁠up:bool -> ?⁠work:Lacaml__.Float64.vec -> ?⁠ofsw:int -> ?⁠w:Lacaml__.Float64.vec -> ?⁠abr:int -> ?⁠abc:int -> Lacaml__.Float64.mat -> Lacaml__.Float64.vec

sbev ?n ?vectors ?zr ?zc ?z ?up ?ofswork ?work ?ofsw ?w ?abr ?abc ab computes all the eigenvalues and, optionally, eigenvectors of the real symmetric band matrix ab.

raises Failure

if the function fails to converge.

returns

the vector w of eigenvalues in ascending order.

raises Failure

if the function fails to converge.

parameter n

default = available number of columns of matrix ab

parameter z

matrix to contain the orthonormal eigenvectors of ab, the i-th column of z holding the eigenvector associated with w.{i}. default = None i.e, eigenvectors are not computed

parameter kd

default = number of rows in matrix ab - 1

parameter up

default = true i.e., upper triangle of the matrix is stored

parameter work

default = vec of minimal length (-> sbev_min_lwork)

parameter ofsw

default = 1 or ignored if w is not given

parameter w

default = vec of length n

parameter abr

default = 1

parameter abc

default = 1

Symmetric-matrix eigenvalue and singular value problems (expert & RRR drivers)

val syevr_min_lwork : int -> int

syevr_min_lwork n

returns

the minimum length of the work-array used by the syevr-function if the logical dimensions of the matrix are n.

val syevr_min_liwork : int -> int

syevr_min_liwork n

returns

the minimum length of the iwork-array used by the syevr-function if the logical dimensions of the matrix are n.

val syevr_opt_lwork : ?⁠n:int -> ?⁠vectors:bool -> ?⁠range:[ `A | `V of float * float | `I of int * int ] -> ?⁠up:bool -> ?⁠abstol:float -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> int

syevr_opt_lwork ?n ?vectors ?range ?up ?abstol ?ar ?ac a

returns

the optimum length of the work-array used by the syevr-function given matrix a, optionally its logical dimension n and whether the eigenvectors must be computed (vectors).

parameter n

default = available number of columns of matrix a

parameter vectors

default = false, i.e. eigenvectors are not computed

parameter up

default = true, i.e. upper triangle of a is stored

val syevr_opt_liwork : ?⁠n:int -> ?⁠vectors:bool -> ?⁠range:[ `A | `V of float * float | `I of int * int ] -> ?⁠up:bool -> ?⁠abstol:float -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> int

syevr_opt_liwork ?n ?vectors ?range ?up ?abstol ?ar ?ac a

returns

the optimum length of the iwork-array used by the syevr-function given matrix a, optionally its logical dimension n and whether the eigenvectors must be computed (vectors).

parameter n

default = available number of columns of matrix a

parameter vectors

default = false, i.e. eigenvectors are not computed

parameter up

default = true, i.e. upper triangle of a is stored

val syevr_opt_l_li_work : ?⁠n:int -> ?⁠vectors:bool -> ?⁠range:[ `A | `V of float * float | `I of int * int ] -> ?⁠up:bool -> ?⁠abstol:float -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> int * int

syevr_opt_l_li_iwork ?n ?vectors ?range ?up ?abstol ?ar ?ac a

returns

the tuple of optimum lengths of the work- and iwork-arrays respectively, used by the syevr-function given matrix a, optionally its logical dimension n and whether the eigenvectors must be computed (vectors).

parameter n

default = available number of columns of matrix a

parameter vectors

default = false, i.e. eigenvectors are not computed

parameter up

default = true, i.e. upper triangle of a is stored

val syevr : ?⁠n:int -> ?⁠vectors:bool -> ?⁠range:[ `A | `V of float * float | `I of int * int ] -> ?⁠up:bool -> ?⁠abstol:float -> ?⁠work:Lacaml__.Float64.vec -> ?⁠iwork:Lacaml.Common.int32_vec -> ?⁠ofsw:int -> ?⁠w:Lacaml__.Float64.vec -> ?⁠zr:int -> ?⁠zc:int -> ?⁠z:Lacaml__.Float64.mat -> ?⁠isuppz:Lacaml.Common.int32_vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> int * Lacaml__.Float64.vec * Lacaml__.Float64.mat * Lacaml.Common.int32_vec

syevr ?n ?vectors ?range ?up ?abstol ?work ?iwork ?ofsw ?w ?zr ?zc ?z ?isuppz ?ar ?ac a range is either `A for computing all eigenpairs, `V (vl, vu) defines the lower and upper range of computed eigenvalues, `I (il, iu) defines the indexes of the computed eigenpairs, which are sorted in ascending order.

returns

the tuple (m, w, z, isuppz), where m is the number of computed eigenpairs, vector w contains the computed eigenvalues in ascending order, z contains the computed eigenvectors in same order, and isuppz indicates the nonzero elements in z.

parameter n

default = available number of columns of matrix a

parameter vectors

default = false i.e, eigenvectors are not computed

parameter range

default = `A

parameter up

default = true i.e., upper triangle of a is stored

parameter abstol

default = result of calling lamch `S

parameter work

default = vec of optimum length (-> syev_opt_lwork)

parameter iwork

default = int32_vec of optimum length (-> syevr_opt_liwork)

parameter ofsw

default = 1 or ignored if w is not given

parameter w

default = vec of length n

parameter zr

default = 1

parameter zc

default = 1

parameter z

default = matrix with minimal required dimension

parameter isuppz

default = int32_vec with minimal required dimension

parameter ar

default = 1

parameter ac

default = 1

val sygv_opt_lwork : ?⁠n:int -> ?⁠vectors:bool -> ?⁠up:bool -> ?⁠itype:[ `A_B | `AB | `BA ] -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> ?⁠br:int -> ?⁠bc:int -> Lacaml__.Float64.mat -> int

sygv_opt_lwork ?n ?vectors ?up ?ar ?ac a ?br ?bc b

returns

the optimum length of the work-array used by the sygv-function for the given matrices a and b, optionally their logical dimension n and whether the eigenvectors must be computed (vectors).

parameter n

default = available number of columns of matrix a

parameter vectors

default = false, i.e. eigenvectors are not computed

parameter up

default = true, i.e. upper triangle of a is stored

parameter itype

specifies the problem type to be solved:

`A_B (default): a*x = (lambda)*a*x

`AB: a*b*x = (lambda)*x

`BA: b*a*x = (lambda)*x

val sygv : ?⁠n:int -> ?⁠vectors:bool -> ?⁠up:bool -> ?⁠work:Lacaml__.Float64.vec -> ?⁠ofsw:int -> ?⁠w:Lacaml__.Float64.vec -> ?⁠itype:[ `A_B | `AB | `BA ] -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> ?⁠br:int -> ?⁠bc:int -> Lacaml__.Float64.mat -> Lacaml__.Float64.vec

sygv ?n ?vectors ?up ?ofswork ?work ?ofsw ?w ?ar ?ac a ?br ?bc b computes all the eigenvalues, and optionally, the eigenvectors of a real generalized symmetric-definite eigenproblem, of the form a*x=(lambda)*b*x, a*b*x=(lambda)*x, or b*a*x=(lambda)*x. Here a and b are assumed to be symmetric and b is also positive definite.

returns

the vector w of eigenvalues in ascending order.

raises Failure

if the function fails to converge.

parameter n

default = available number of columns of matrix a

parameter vectors

default = false i.e, eigenvectors are not computed

parameter up

default = true i.e., upper triangle of a is stored

parameter work

default = vec of optimum length (-> sygv_opt_lwork)

parameter ofsw

default = 1 or ignored if w is not given

parameter w

default = vec of length n

parameter itype

specifies the problem type to be solved:

`A_B (default): a*x = (lambda)*a*x

`AB: a*b*x = (lambda)*x

`BA: b*a*x = (lambda)*x

val sbgv : ?⁠n:int -> ?⁠ka:int -> ?⁠kb:int -> ?⁠zr:int -> ?⁠zc:int -> ?⁠z:Lacaml__.Float64.mat -> ?⁠up:bool -> ?⁠work:Lacaml__.Float64.vec -> ?⁠ofsw:int -> ?⁠w:Lacaml__.Float64.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> ?⁠br:int -> ?⁠bc:int -> Lacaml__.Float64.mat -> Lacaml__.Float64.vec

sbgv ?n ?ka ?kb ?zr ?zc ?z ?up ?work ?ofsw ?w ?ar ?ac a ?br ?bc b computes all the eigenvalues, and optionally, the eigenvectors of a real generalized symmetric-definite banded eigenproblem, of the form a*x=(lambda)*b*x. Here a and b are assumed to be symmetric and banded, and b is also positive definite.

returns

the vector w of eigenvalues in ascending order.

raises Failure

if the function fails to converge.

parameter n

default = available number of columns of matrix a

parameter ka

the number of superdiagonals (or subdiagonals if up = false) of the matrix a. Default = dim1 a - ar.

parameter kb

same as ka but for the matrix b.

parameter z

default = None i.e, eigenvectors are not computed

parameter up

default = true i.e., upper triangle of a is stored

parameter work

default = vec of optimum length (3 * n)

parameter ofsw

default = 1 or ignored if w is not given

parameter w

default = vec of length n