Module Lacaml__Impl2_C

BLAS-1 interface

val dotu : ?⁠n:int -> ?⁠ofsx:int -> ?⁠incx:int -> Lacaml__.Complex32.vec -> ?⁠ofsy:int -> ?⁠incy:int -> Lacaml__.Complex32.vec -> Stdlib.Complex.t

dotu ?n ?ofsx ?incx x ?ofsy ?incy y see BLAS documentation!

parameter n

default = greater n s.t. ofsx+(n-1)(abs incx) <= dim x

parameter ofsx

default = 1

parameter incx

default = 1

parameter ofsy

default = 1

parameter incy

default = 1

val dotc : ?⁠n:int -> ?⁠ofsx:int -> ?⁠incx:int -> Lacaml__.Complex32.vec -> ?⁠ofsy:int -> ?⁠incy:int -> Lacaml__.Complex32.vec -> Stdlib.Complex.t

dotc ?n ?ofsx ?incx x ?ofsy ?incy y see BLAS documentation!

parameter n

default = greater n s.t. ofsx+(n-1)(abs incx) <= dim x

parameter ofsx

default = 1

parameter incx

default = 1

parameter ofsy

default = 1

parameter incy

default = 1

LAPACK interface

val lansy_min_lwork : int -> Lacaml.Common.norm4 -> int

lansy_min_lwork m norm

returns

the minimum length of the work array used by the lansy-function.

parameter norm

type of norm that will be computed by lansy

parameter n

the number of columns (and rows) in the matrix

val lansy : ?⁠n:int -> ?⁠up:bool -> ?⁠norm:Lacaml.Common.norm4 -> ?⁠work:Lacaml__.Complex32.rvec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Complex32.mat -> float

lansy ?n ?up ?norm ?work ?ar ?ac a see LAPACK documentation!

parameter norm

default = `O

parameter up

default = true (reference upper triangular part of a)

parameter n

default = number of columns of matrix a

parameter work

default = allocated work space for norm `I

val gecon_min_lwork : int -> int

gecon_min_lwork n

returns

the minimum length of the work array used by the gecon-function.

parameter n

the logical dimensions of the matrix given to the gecon-function

val gecon_min_lrwork : int -> int

gecon_min_lrwork n

returns

the minimum length of the rwork array used by the gecon-function.

parameter n

the logical dimensions of the matrix given to gecon-function

val gecon : ?⁠n:int -> ?⁠norm:Lacaml.Common.norm2 -> ?⁠anorm:float -> ?⁠work:Lacaml__.Complex32.vec -> ?⁠rwork:Lacaml__.Complex32.rvec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Complex32.mat -> float

gecon ?n ?norm ?anorm ?work ?rwork ?ar ?ac a

returns

estimate of the reciprocal of the condition number of matrix a

parameter n

default = available number of columns of matrix a

parameter norm

default = 1-norm

parameter anorm

default = norm of the matrix a as returned by lange

parameter work

default = automatically allocated workspace

parameter rwork

default = automatically allocated workspace

parameter ar

default = 1

parameter ac

default = 1

val sycon_min_lwork : int -> int

sycon_min_lwork n

returns

the minimum length of the work array used by the sycon-function.

parameter n

the logical dimensions of the matrix given to the sycon-function

val sycon : ?⁠n:int -> ?⁠up:bool -> ?⁠ipiv:Lacaml.Common.int32_vec -> ?⁠anorm:float -> ?⁠work:Lacaml__.Complex32.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Complex32.mat -> float

sycon ?n ?up ?ipiv ?anorm ?work ?ar ?ac a

returns

estimate of the reciprocal of the condition number of symmetric matrix a

parameter n

default = available number of columns of matrix a

parameter up

default = upper triangle of the factorization of a is stored

parameter ipiv

default = vec of length n

parameter anorm

default = 1-norm of the matrix a as returned by lange

parameter work

default = automatically allocated workspace

val pocon_min_lwork : int -> int

pocon_min_lwork n

returns

the minimum length of the work array used by the pocon-function.

parameter n

the logical dimensions of the matrix given to the pocon-function

val pocon_min_lrwork : int -> int

pocon_min_lrwork n

returns

the minimum length of the rwork array used by the pocon-function.

parameter n

the logical dimensions of the matrix given to pocon-function

val pocon : ?⁠n:int -> ?⁠up:bool -> ?⁠anorm:float -> ?⁠work:Lacaml__.Complex32.vec -> ?⁠rwork:Lacaml__.Complex32.rvec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Complex32.mat -> float

pocon ?n ?up ?anorm ?work ?rwork ?ar ?ac a

returns

estimate of the reciprocal of the condition number of complex Hermitian positive definite matrix a

parameter n

default = available number of columns of matrix a

parameter up

default = upper triangle of Cholesky factorization of a is stored

parameter work

default = automatically allocated workspace

parameter rwork

default = automatically allocated workspace

parameter anorm

default = 1-norm of the matrix a as returned by lange

General Schur factorization

val gees : ?⁠n:int -> ?⁠jobvs:Lacaml.Common.schur_vectors -> ?⁠sort:Lacaml.Common.eigen_value_sort -> ?⁠w:Lacaml__.Complex32.vec -> ?⁠vsr:int -> ?⁠vsc:int -> ?⁠vs:Lacaml__.Complex32.mat -> ?⁠work:Lacaml__.Complex32.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Complex32.mat -> int * Lacaml__.Complex32.vec * Lacaml__.Complex32.mat

gees ?n ?jobvs ?sort ?w ?vsr ?vsc ?vs ?work ?ar ?ac a See gees-function for details about arguments.

returns

(sdim, w, vs)

General SVD routines

val gesvd_min_lwork : m:int -> n:int -> int

gesvd_min_lwork ~m ~n

returns

the minimum length of the work array used by the gesvd-function for matrices with m rows and n columns.

val gesvd_lrwork : m:int -> n:int -> int

gesvd_lrwork m n

returns

the (minimum) length of the rwork array used by the gesvd-function.

val gesvd_opt_lwork : ?⁠m:int -> ?⁠n:int -> ?⁠jobu:Lacaml.Common.svd_job -> ?⁠jobvt:Lacaml.Common.svd_job -> ?⁠s:Lacaml__.Complex32.rvec -> ?⁠ur:int -> ?⁠uc:int -> ?⁠u:Lacaml__.Complex32.mat -> ?⁠vtr:int -> ?⁠vtc:int -> ?⁠vt:Lacaml__.Complex32.mat -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Complex32.mat -> int

val gesvd : ?⁠m:int -> ?⁠n:int -> ?⁠jobu:Lacaml.Common.svd_job -> ?⁠jobvt:Lacaml.Common.svd_job -> ?⁠s:Lacaml__.Complex32.rvec -> ?⁠ur:int -> ?⁠uc:int -> ?⁠u:Lacaml__.Complex32.mat -> ?⁠vtr:int -> ?⁠vtc:int -> ?⁠vt:Lacaml__.Complex32.mat -> ?⁠work:Lacaml__.Complex32.vec -> ?⁠rwork:Lacaml__.Complex32.rvec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Complex32.mat -> Lacaml__.Complex32.rvec * Lacaml__.Complex32.mat * Lacaml__.Complex32.mat

General eigenvalue problem (simple drivers)

val geev_min_lwork : int -> int

geev_min_lwork n

returns

the minimum length of the work array used by the geev-function.

parameter n

the logical dimensions of the matrix given to geev-function

val geev_min_lrwork : int -> int

geev_min_lrwork n

returns

the minimum length of the rwork array used by the geev-function.

parameter n

the logical dimensions of the matrix given to geev-function

val geev_opt_lwork : ?⁠n:int -> ?⁠vlr:int -> ?⁠vlc:int -> ?⁠vl:Lacaml__.Complex32.mat option -> ?⁠vrr:int -> ?⁠vrc:int -> ?⁠vr:Lacaml__.Complex32.mat option -> ?⁠ofsw:int -> ?⁠w:Lacaml__.Complex32.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Complex32.mat -> int

geev ?work ?rwork ?n ?vlr ?vlc ?vl ?vrr ?vrc ?vr ?ofsw w ?ar ?ac a See geev-function for details about arguments.

returns

"optimal" work size

val geev : ?⁠n:int -> ?⁠work:Lacaml__.Complex32.vec -> ?⁠rwork:Lacaml__.Complex32.vec -> ?⁠vlr:int -> ?⁠vlc:int -> ?⁠vl:Lacaml__.Complex32.mat option -> ?⁠vrr:int -> ?⁠vrc:int -> ?⁠vr:Lacaml__.Complex32.mat option -> ?⁠ofsw:int -> ?⁠w:Lacaml__.Complex32.vec -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Complex32.mat -> Lacaml__.Complex32.mat * Lacaml__.Complex32.vec * Lacaml__.Complex32.mat

geev ?work ?rwork ?n ?vlr ?vlc ?vl ?vrr ?vrc ?vr ?ofsw w ?ar ?ac a

returns

(lv, w, rv), where lv and rv correspond to the left and right eigenvectors respectively, w to the eigenvalues. lv (rv) is the empty matrix if vl (vr) is set to None.

raises Failure

if the function fails to converge

parameter n

default = available number of columns of matrix a

parameter work

default = automatically allocated workspace

parameter rwork

default = automatically allocated workspace

parameter vl

default = Automatically allocated left eigenvectors. Pass None if you do not want to compute them, Some lv if you want to provide the storage. You can set vlr, vlc in the last case. (See LAPACK GEEV docs for details about storage of complex eigenvectors)

parameter vr

default = Automatically allocated right eigenvectors. Pass None if you do not want to compute them, Some rv if you want to provide the storage. You can set vrr, vrc in the last case.

parameter w

default = automatically allocate eigenvalues

parameter a

the matrix whose eigensystem is computed