Module Lacaml__Mat2_D

Matrix operations

Creation of matrices

val hilbert : int -> Lacaml__.Float64.mat

hilbert n

returns

an nxn Hilbert matrix.

val hankel : int -> Lacaml__.Float64.mat

hankel n

returns

an nxn Hankel matrix.

val pascal : int -> Lacaml__.Float64.mat

pascal n

returns

an nxn Pascal matrix.

val rosser : unit -> Lacaml__.Float64.mat

rosser n

returns

8x8 Rosser matrix.

val toeplitz : Lacaml__.Float64.vec -> Lacaml__.Float64.mat

toeplitz v

returns

the Toeplitz matrix associated with v. The constant diagonals are read from left to right from v.

raises Invalid_argument

if the length of v is not an odd number.

val vandermonde : Lacaml__.Float64.vec -> Lacaml__.Float64.mat

vandermonde v

returns

the Vandermonde matrix associated with v.

val wilkinson : int -> Lacaml__.Float64.mat

wilkinson n

returns

the nxn Wilkinson matrix.

raises Invalid_argument

if n is not an odd number >= 3.

val random : ?⁠rnd_state:Stdlib.Random.State.t -> ?⁠from:float -> ?⁠range:float -> int -> int -> Lacaml__.Float64.mat

random ?rnd_state ?from ?range m n

returns

an mxn matrix initialized with random elements sampled uniformly from range starting at from. A random state rnd_state can be passed.

parameter rnd_state

default = Random.get_state ()

parameter from

default = -1.0

parameter range

default = 2.0

Unary matrix operations

val abs : Lacaml__.Float64.Types.Mat.unop

abs ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the absolute value of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val signum : Lacaml__.Float64.Types.Mat.unop

signum ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the sign value (-1 for negative numbers, 0 (or -0) for zero, 1 for positive numbers, nan for nan) of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val sqr : Lacaml__.Float64.Types.Mat.unop

sqr ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the square of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val sqrt : Lacaml__.Float64.Types.Mat.unop

sqrt ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the square root of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val cbrt : Lacaml__.Float64.Types.Mat.unop

cbrt ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the cubic root of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val exp : Lacaml__.Float64.Types.Mat.unop

exp ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the exponential of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val exp2 : Lacaml__.Float64.Types.Mat.unop

exp2 ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the base-2 exponential of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val expm1 : Lacaml__.Float64.Types.Mat.unop

expm1 ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes exp a -. 1. of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val log : Lacaml__.Float64.Types.Mat.unop

log ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the logarithm of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val log10 : Lacaml__.Float64.Types.Mat.unop

log10 ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the base-10 logarithm of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val log2 : Lacaml__.Float64.Types.Mat.unop

log2 ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes base-2 logarithm of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val log1p : Lacaml__.Float64.Types.Mat.unop

log1p ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes log (1 + a) of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val sin : Lacaml__.Float64.Types.Mat.unop

sin ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the sine of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val cos : Lacaml__.Float64.Types.Mat.unop

cos ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the cosine of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val tan : Lacaml__.Float64.Types.Mat.unop

tan ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the tangent of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val asin : Lacaml__.Float64.Types.Mat.unop

asin ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the arc sine of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val acos : Lacaml__.Float64.Types.Mat.unop

acos ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the arc cosine of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val atan : Lacaml__.Float64.Types.Mat.unop

atan ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the arc tangent of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val sinh : Lacaml__.Float64.Types.Mat.unop

sinh ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the hyperbolic sine of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val cosh : Lacaml__.Float64.Types.Mat.unop

cosh ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the hyperbolic cosine of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val tanh : Lacaml__.Float64.Types.Mat.unop

tanh ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the hyperbolic tangent of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val asinh : Lacaml__.Float64.Types.Mat.unop

asinh ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the hyperbolic arc sine of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val acosh : Lacaml__.Float64.Types.Mat.unop

acosh ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the hyperbolic arc cosine of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val atanh : Lacaml__.Float64.Types.Mat.unop

atanh ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the hyperbolic arc tangent of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val floor : Lacaml__.Float64.Types.Mat.unop

floor ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the floor of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val ceil : Lacaml__.Float64.Types.Mat.unop

ceil ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the ceiling of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val round : Lacaml__.Float64.Types.Mat.unop

round ?patt ?m ?n ?br ?bc ?b ?ar ?ac a rounds the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val trunc : Lacaml__.Float64.Types.Mat.unop

trunc ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the truncation of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val erf : Lacaml__.Float64.Types.Mat.unop

erf ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the error function of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val erfc : Lacaml__.Float64.Types.Mat.unop

erfc ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the complementary error function of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val logistic : Lacaml__.Float64.Types.Mat.unop

logistic ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the logistic function 1/(1 + exp(-a) of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val relu : Lacaml__.Float64.Types.Mat.unop

relu ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the rectified linear unit function max(a, 0) of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val softplus : Lacaml__.Float64.Types.Mat.unop

softplus ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the softplus function log(1 + exp(x) of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

val softsign : Lacaml__.Float64.Types.Mat.unop

softsign ?patt ?m ?n ?br ?bc ?b ?ar ?ac a computes the softsign function x / (1 + abs(x)) of the elements in the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt. If b is given, the result will be stored in there using offsets br and bc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter br

default = 1

parameter bc

default = 1

parameter b

default = fresh matrix with br + m - 1 rows and bc + n - 1 columns

parameter ar

default = 1

parameter ac

default = 1

Binary matrix operations

val pow : Lacaml__.Float64.Types.Mat.binop

pow ?patt ?m ?n ?cr ?cc ?c ?ar ?ac a ?br ?bc b computes pow(a, b) for the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt with the corresponding sub-matrix of the matrix b starting in row br and column bc. If c is given, the result will be stored in there starting in row cr and column cc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter cr

default = 1

parameter cc

default = 1

parameter c

default = fresh matrix with cr + m - 1 rows and cc + n - 1 columns

parameter br

default = 1

parameter bc

default = 1

parameter ar

default = 1

parameter ac

default = 1

val atan2 : Lacaml__.Float64.Types.Mat.binop

atan2 ?patt ?m ?n ?cr ?cc ?c ?ar ?ac a ?br ?bc b computes atan2(a, b) for the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt with the corresponding sub-matrix of the matrix b starting in row br and column bc. If c is given, the result will be stored in there starting in row cr and column cc, otherwise a fresh matrix will be used. The resulting matrix is returned.

NOTE: WARNING! From a geometric point of view, the atan2 function takes the y-coordinate in a and the x-coordinate in b. This confusion is a sad consequence of the C99-standard reversing the argument order for atan2 for no good reason.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter cr

default = 1

parameter cc

default = 1

parameter c

default = fresh matrix with cr + m - 1 rows and cc + n - 1 columns

parameter br

default = 1

parameter bc

default = 1

parameter ar

default = 1

parameter ac

default = 1

val hypot : Lacaml__.Float64.Types.Mat.binop

hypot ?patt ?m ?n ?cr ?cc ?c ?ar ?ac a ?br ?bc b computes sqrt(a*a+b*b) for the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt with the corresponding sub-matrix of the matrix b starting in row br and column bc. If c is given, the result will be stored in there starting in row cr and column cc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter cr

default = 1

parameter cc

default = 1

parameter c

default = fresh matrix with cr + m - 1 rows and cc + n - 1 columns

parameter br

default = 1

parameter bc

default = 1

parameter ar

default = 1

parameter ac

default = 1

val min2 : Lacaml__.Float64.Types.Mat.binop

min2 ?patt ?m ?n ?cr ?cc ?c ?ar ?ac a ?br ?bc b computes the elementwise minimum of the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt with the corresponding sub-matrix of the matrix b starting in row br and column bc. If c is given, the result will be stored in there starting in row cr and column cc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter cr

default = 1

parameter cc

default = 1

parameter c

default = fresh matrix with cr + m - 1 rows and cc + n - 1 columns

parameter br

default = 1

parameter bc

default = 1

parameter ar

default = 1

parameter ac

default = 1

val max2 : Lacaml__.Float64.Types.Mat.binop

max2 ?patt ?m ?n ?cr ?cc ?c ?ar ?ac a ?br ?bc b computes the elementwise maximum of the m by n sub-matrix of the matrix a starting in row ar and column ac and pattern patt with the corresponding sub-matrix of the matrix b starting in row br and column bc. If c is given, the result will be stored in there starting in row cr and column cc, otherwise a fresh matrix will be used. The resulting matrix is returned.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter cr

default = 1

parameter cc

default = 1

parameter c

default = fresh matrix with cr + m - 1 rows and cc + n - 1 columns

parameter br

default = 1

parameter bc

default = 1

parameter ar

default = 1

parameter ac

default = 1

val sum_prod : ?⁠patt:Lacaml__.Float64.Types.Mat.patt -> ?⁠m:int -> ?⁠n:int -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> ?⁠br:int -> ?⁠bc:int -> Lacaml__.Float64.mat -> float

sum_prod ?patt ?m ?n ?ar ?ac a ?br ?bc b

returns

the sum of elementwise products between the m by n sub-matrix of the matrix a starting in row ar and column ac with the corresponding sub-matrix of the matrix b starting in row br and column bc.

parameter patt

default = `Full

parameter m

default = greater n s.t. ar + m - 1 <= dim1 a

parameter n

default = greater n s.t. ac + n - 1 <= dim2 a

parameter ar

default = 1

parameter ac

default = 1

parameter br

default = 1

parameter bc

default = 1

Miscellaneous functions

val log_sum_exp : ?⁠patt:Lacaml__.Float64.Types.Mat.patt -> ?⁠m:int -> ?⁠n:int -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> float

log_sum_exp ?patt ?m ?n ?ar ?ac a computes the logarithm of the sum of exponentials of all elements in the m-by-n submatrix using pattern patt, starting at row ar and column ac.

parameter patt

default = `Full

parameter m

default = number of rows of a

parameter n

default = number of columns of a

Ternary matrix operations

val cpab : ?⁠patt:Lacaml__.Float64.Types.Mat.patt -> ?⁠m:int -> ?⁠n:int -> ?⁠cr:int -> ?⁠cc:int -> Lacaml__.Float64.mat -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> ?⁠br:int -> ?⁠bc:int -> Lacaml__.Float64.mat -> unit

cpab ?patt ?m ?n ?cr ?cc c ?ar ?ac a ?br ?bc b multiplies designated m-by-n range of elements of matrices a and b using pattern patt elementwise and adds the result to and stores it in the specified range in c. This function is useful for convolutions. Similar to Vec.zpxy.

parameter patt

default = `Full

parameter m

default = number of rows of a

parameter n

default = number of columns of a

parameter cr

default = 1

parameter cc

default = 1

parameter ar

default = 1

parameter ac

default = 1

parameter br

default = 1

parameter bc

default = 1

val cmab : ?⁠patt:Lacaml__.Float64.Types.Mat.patt -> ?⁠m:int -> ?⁠n:int -> ?⁠cr:int -> ?⁠cc:int -> Lacaml__.Float64.mat -> ?⁠ar:int -> ?⁠ac:int -> Lacaml__.Float64.mat -> ?⁠br:int -> ?⁠bc:int -> Lacaml__.Float64.mat -> unit

cmab ?patt ?m ?n ?cr ?cc c ?ar ?ac a ?br ?bc b multiplies designated m-by-n range of elements of matrices a and b elementwise using pattern patt and subtracts the result from and stores it in the specified range in c. This function is useful for convolutions. Similar to Vec.zmxy.

parameter patt

default = `Full

parameter m

default = number of rows of a

parameter n

default = number of columns of a

parameter cr

default = 1

parameter cc

default = 1

parameter ar

default = 1

parameter ac

default = 1

parameter br

default = 1

parameter bc

default = 1